// src/app/models/types/other-document-upload.ts
export interface OtherDocumentUpload {}

export interface OtherDocument {
    IsUser: boolean;
    DocumentName: string;
    DocumentType: string;
    UploadUrl: string;
    UserId: string;
  }

  export interface OtherDocumentCompany {
    IsUser: boolean;
    DocumentName: string;
    DocumentType: string;
    UploadUrl: string;
    CompanyId: string;
  }
    